add stack chart
