+++ src/components/area-chart.tsx
@@ -1,160 +1,179 @@
-import React, { useEffect } from "react";
-import useMediaQuery from "@/hooks/useMediaQuery";
-import { Line } from "react-chartjs-2";
-import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Title,
-  Tooltip,
-  Filler,
-  Legend,
-  Decimation,
-  TimeScale,
-} from "chart.js";
+import * as Highcharts from "highcharts/highstock";
+import HighchartsReact from "highcharts-react-official";
+import { Dispatch, SetStateAction, useEffect, useRef } from "react";
+import { useState } from "react";
+import { useViewportSize } from "@mantine/hooks";
 
-import "chartjs-adapter-luxon";
-
-export interface AreaChartProps {
-  title: string;
-  labels: string[] | string[][] | number[];
-  datasets: DatasetOptions[];
-  callbacks?: {
-    label?: (context: any) => string;
+export interface InitialAreaChatStateInput {
+  stacking?: boolean; // Custom option
+  title: {
+    text: string;
   };
-  yAxisTicksCallback?: (value: number, index: any, values: any) => string;
-  stacked?: boolean;
+  yAxis: {
+    labels: {
+      formatter: Highcharts.AxisLabelsFormatterCallbackFunction;
+    };
+    max?: number;
+  };
+  tooltip: {
+    pointFormatter: Highcharts.FormatterCallbackFunction<Highcharts.Point>;
+  };
 }
 
-export interface DatasetOptions {
-  label: string;
-  fill: boolean;
-  data: number[] | any[];
-  backgroundColor?: string;
-  borderColor?: string;
-  yAxisId: string;
-}
-
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Title,
-  Tooltip,
-  Filler,
-  Legend,
-  Decimation,
-  TimeScale
-);
-
-const AreaChart: React.FC<AreaChartProps> = ({
-  labels,
-  datasets,
-  title,
-  stacked = false,
-  callbacks,
-  yAxisTicksCallback,
-}) => {
-  const { isSmallerThanMd, isLargerThanXl, isSmallerThanXs } = useMediaQuery();
-
-  const maxTicksLimit = isLargerThanXl || (isSmallerThanMd && !isSmallerThanXs) ? 8 : 4;
-
-  const chartData = {
-    labels,
-    datasets: datasets.map((dataset) => dataset),
-  };
-
-  const options = {
-    responsive: true,
-    maintainAspectRatio: false,
-    parsing: false,
-    animation: false,
-    spanGaps: true,
-    tooltip: {
-      mode: "index",
-    },
-    interaction: {
-      mode: "nearest",
-      axis: "x",
-      intersect: false,
-    },
-    elements: {
-      point: {
-        borderWidth: 0,
-        radius: 0, // Makes points invisible unless hovered
-        backgroundColor: "rgba(0,0,0,0)",
-      },
-    },
-    plugins: {
-      tooltip: {
-        mode: "index",
-        callbacks,
-      },
-      decimation: {
-        enabled: true,
-        algorithm: "min-max" as const,
-        samples: 60,
-        threshold: 60,
-      },
-      legend: {
-        position: "top" as const,
-        labels: {
-          color: "#8192ac",
-        },
-      },
-      title: {
-        display: true,
-        text: title,
+export const useAreaChartState = (
+  opts: InitialAreaChatStateInput
+): [Highcharts.Options, Dispatch<SetStateAction<Highcharts.Options>>] => {
+  const [chartOptions, setChartOptions] = useState<Highcharts.Options>({
+    title: {
+      text: opts.title.text,
+      style: {
+        fontFamily: "Roboto, Arial, sans-serif",
+        fontWeight: "bold",
+        fontSize: "18px",
         color: "#dce1e8",
-        align: "start" as const,
-        font: {
-          family: "Roboto, Arial, sans-serif",
-          weight: "bold",
-          size: 18,
+      },
+    },
+    plotOptions: {
+      series: {
+        marker: {
+          enabled: false,
+        },
+      },
+      area: {
+        stacking: "normal",
+        // lineColor: "#666666",
+        lineWidth: 1,
+        marker: {
+          lineWidth: 1,
+          // lineColor: "#666666",
         },
       },
     },
-    scales: {
-      y: {
-        min: 0,
-        stacked,
-        ticks: {
+    xAxis: {
+      type: "datetime",
+      gridLineColor: "#263858",
+      lineColor: "#263858",
+      labels: {
+        step: 2,
+        format: "{value:%I:%M %p}",
+        style: {
           color: "#8192ac",
-          callback: yAxisTicksCallback,
-        },
-        grid: {
-          color: "#263858",
-        },
-      },
-      x: {
-        min: undefined,
-        type: "time",
-        time: {
-          round: "seconds",
-          displayFormats: {
-            millisecond: "hh:mm a",
-            second: "hh:mm a",
-          },
-        },
-
-        ticks: {
-          color: "#8192ac",
-          maxTicksLimit: maxTicksLimit,
-          autoSkip: true,
-          maxRotation: 0,
-          source: "auto",
-          align: "end",
-        },
-        grid: {
-          color: "#263858",
         },
       },
     },
-  };
+    legend: {
+      enabled: false,
+    },
+    time: {
+      useUTC: false,
+    },
 
-  return <Line data={chartData as any} options={options as any} />;
+    yAxis: {
+      max: opts.yAxis.max,
+      title: {
+        text: null,
+      },
+      startOnTick: true,
+      gridLineColor: "#263858",
+      lineColor: "#263858",
+      labels: {
+        formatter: opts.yAxis.labels.formatter,
+        style: {
+          color: "white",
+        },
+      },
+    },
+    credits: {
+      enabled: false,
+    },
+    tooltip: {
+      xDateFormat: "%Y-%m-%d %I:%M:%S %p",
+      pointFormatter: opts.tooltip.pointFormatter,
+      style: {
+        color: "#dce1e8",
+      },
+      backgroundColor: "#263858",
+    },
+    scrollbar: {
+      rifleColor: "#324363",
+      barBackgroundColor: "#324363",
+      buttonBackgroundColor: "#324363",
+      trackBorderColor: "#324363",
+    },
+    rangeSelector: {
+      labelStyle: {
+        color: "#8192ac",
+        backgroundColor: "#263858",
+      },
+
+      inputStyle: {
+        color: "#8192ac",
+      },
+      buttonTheme: {
+        fill: "none",
+        display: "none",
+        r: 8,
+        style: {
+          background: "none",
+          color: "#8192ac",
+          backgroundColor: "#263858",
+          fontWeight: "bold",
+        },
+      },
+      buttons: [
+        {
+          count: 1,
+          type: "minute",
+          text: "1M",
+        },
+        {
+          count: 5,
+          type: "minute",
+          text: "5M",
+        },
+        {
+          count: 30,
+          type: "minute",
+          text: "30M",
+        },
+        {
+          type: "all",
+          text: "All",
+        },
+      ],
+      // inputEnabled: false,
+      selected: 0,
+    },
+
+    chart: {
+      backgroundColor: "transparent",
+      style: {
+        color: "#dce1e8",
+      },
+    },
+  });
+
+  return [chartOptions, setChartOptions];
 };
+
+const AreaChart: React.FC<HighchartsReact.Props> = (props) => {
+  const chartComponentRef = useRef<HighchartsReact.RefObject>(null);
+  const { width } = useViewportSize();
+
+  // Reflow chart on window resize
+  useEffect(() => {
+    chartComponentRef.current?.chart?.reflow();
+  }, [width]);
+
+  return (
+    <HighchartsReact
+      highcharts={Highcharts}
+      options={props.options}
+      ref={chartComponentRef}
+      constructorType={"stockChart"}
+      containerProps={{ style: { height: "100%", width: "100%" } }}
+    />
+  );
+};
+
 export default AreaChart;
