+++ src/features/metrics/components/global-cpu/global-cpu.area-chart.tsx
@@ -1,6 +1,6 @@
 import Card from "@/components/card";
 
-import AreaChart, { useAreaChartState } from "@/components/area-chart.prototype";
+import AreaChart, { useAreaChartState } from "@/components/area-chart";
 import useServerEventsContext from "@/hooks/useServerEventsContext";
 import { useEffect } from "react";
